+++ HISTORY.md
@@ -1,5 +1,15 @@
 # History
 
+## 1.2.0
+### Added
+- File : A class used to represent a file
+- Csv : A class used to represent a csv file
+- Sas : A class used to represent a sas7dbat file
+- Excel : A class used to represent a xlsx file
+### Fixed
+- Database : A class used to represent a database.
+### Deprecated
+
 ## 0.1.0
 ### Added
 - DbConnInfoStore : A class used to represent a store that contains databases connection information.